backend/requirements.txt
langchain
langchain-openai
chromadb
python-docx
pytesseract
pandas
openai
uvicorn
fastapi
langchain_community
openpyxl
rapidfuzz
python-dotenv
langchain-experimental
langchain_chroma
rank_bm25
docling
scikit-learn 
numpy
google-generativeai